+++ keyboards/kinesis/alvicstep/rules.mk
@@ -1,0 +1,4 @@
+CUSTOM_MATRIX = yes # need to do our own thing with the matrix
+
+# Project specific files
+SRC += matrix.c
