+++ keyboards/kinesis/rules.mk
@@ -28,9 +28,5 @@ MIDI_ENABLE = no            # MIDI controls
 UNICODE_ENABLE = no         # Unicode
 BLUETOOTH_ENABLE = no       # Enable Bluetooth with the Adafruit EZ-Key HID
 AUDIO_ENABLE = no # Audio output should be port E6, current quantum library hardcodes C6, which we use for programming
-CUSTOM_MATRIX=yes # need to do our own thing with the matrix
 
 DEFAULT_FOLDER = kinesis/alvicstep
-
-# Project specific files
-SRC = matrix.c
