--- quantum/split_common/matrix.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <common/matrix.h>
